+++ index.js
@@ -1,5 +1,7 @@
 let _inventory = require('@architect/inventory')
 let { updater } = require('@architect/utils')
+let cleanup = require('./src/utils/cleanup')
+
 let direct = require('./src/direct')
 let sam = require('./src/sam')
 let _static = require('./src/static')
@@ -16,8 +18,15 @@ function run (mod) {
       })
     }
 
+    // Always attempt to clean up after ourselves before exiting
+    function clean (err, result) {
+      cleanup()
+      if (err) callback(err)
+      else callback(null, result)
+    }
+
     // Entered via CLI (or something that supplied inventory)
-    if (options.inventory) mod(options, callback)
+    if (options.inventory) mod(options, clean)
     else {
       // Get inventory, but don't fetch env vars if it's a dry-run
       _inventory({ env: true }, function (err, inv) {
@@ -26,7 +35,7 @@ function run (mod) {
           options.update = updater('Deploy')
           options.region = options.region || inv.inv.aws.region
           options.inventory = inv
-          mod(options, callback)
+          mod(options, clean)
         }
       })
     }