+++ src/sam/02-after/05-cleanup.js
@@ -1,7 +1,0 @@
-let { join } = require('path')
-let rm = require('rimraf')
-
-module.exports = function cleanup (callback) {
-  // Clean up temp dir from root proxy + fingerprint
-  rm(join(process.cwd(), '__ARC_TMP__'), callback)
-}